Address review: extract AlertMixin for re-use
(cherry picked from commit 2e826763ce3b0206c20f1bd5b7f11c5d6b314b81)
